requirements.txt
requests==2.*
selenium >= 3.11.0,<=4.8.3
percy-selenium==2.1.2
